Restructure and begin to document.
Have added a new index page to list the various mockups and their purpose and starting to document
the python server to make use of that. This required the data links in the example JS to be updated
to reflect the new server root location (not sure why the relative paths didn't work). Also adding
a bit more work on the example.